+++ Applications/Quickstarts.Servers/MemoryBuffer/MemoryBufferMonitoredItem.cs
@@ -46,7 +46,7 @@ namespace MemoryBuffer
         public MemoryBufferMonitoredItem(
             IServerInternal server,
             INodeManager nodeManager,
-            object mangerHandle,
+            object managerHandle,
             uint offset,
             uint subscriptionId,
             uint id,
@@ -67,7 +67,7 @@ namespace MemoryBuffer
             base(
                 server,
                 nodeManager,
-                mangerHandle,
+                managerHandle,
                 subscriptionId,
                 id,
                 itemToMonitor,
@@ -93,14 +93,14 @@ namespace MemoryBuffer
         public MemoryBufferMonitoredItem(
             IServerInternal server,
             INodeManager nodeManager,
-            object mangerHandle,
+            object managerHandle,
             uint offset,
             IStoredMonitoredItem storedMonitoredItem)
         :
             base(
                 server,
                 nodeManager,
-                mangerHandle,
+                managerHandle,
                 storedMonitoredItem)
         {
             m_offset = offset;
