+++ Applications/Quickstarts.Servers/ReferenceServer/ReferenceServer.cs
@@ -95,7 +95,7 @@ namespace Quickstarts.ReferenceServer
         /// </summary>
         /// <param name="server">The server.</param>
         /// <param name="configuration">The configuration.</param>
-        /// <returns>Returns a subscriptionStoe for a server, the return type is <seealso cref="ISubscriptionStore"/>.</returns>
+        /// <returns>Returns a subscriptionStore for a server, the return type is <seealso cref="ISubscriptionStore"/>.</returns>
         protected override ISubscriptionStore CreateSubscriptionStore(IServerInternal server, ApplicationConfiguration configuration)
         {
             if (configuration?.ServerConfiguration?.DurableSubscriptionsEnabled == true)
